Create trade_game_2
This is the second version of trade game. Now it allows you to have any number of stocks (add the ticker into line 5). And because there's more choice a limit has put on the amount of trades allowed a day to limit transaction fees (this is toggle-able in line 21 with True and False).