Create Mew Bot
This bot takes a screen shot of your main monitor thats running pokemon showdown. then tells you the type effectivity of all pokemon on screen. yes you can type /weakness "pokemon_name" and it'll give it to you, but this is faster, does it for all, and you don't have to type anything!